custom-snippets/test-create-table/requirements.txt
flask_sqlalchemy
psycopg2
pytest
pytest-flask-sqlalchemy